screenpipe-admin/backend/pyproject.toml
[tool.poetry]
name = "timeglass-backend"
version = "0.1.0"
description = "TimeGlass Backend Service"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = "^0.23.2"
elasticsearch = "^8.9.0"
python-dotenv = "^1.0.0"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
requests = "^2.31.0"
sqlalchemy = "^2.0.38"
pymysql = "^1.1.1"
alembic = "^1.14.1"
aiomysql = "^0.2.0"
greenlet = "^3.1.1"
websockets = "^12.0.0"
starlette = "^0.27.0"
python-multipart = "^0.0.20"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
httpx = "^0.25.1"
black = "^23.10.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
